Struct cdchunking::ChunkStream

pub struct ChunkStream<R: Read, I: ChunkerImpl> { /* fields omitted */ }

Methods

impl<R: Read, I: ChunkerImpl> ChunkStream<R, I>

pub fn read<'a>(&'a mut self) -> Option<Result<ChunkInput<'a>>>

Iterate on the chunks, returning ChunkInput items.

An item is either some data that is part of the current chunk, or End, indicating the boundary between chunks.

End is always returned at the end of the last chunk.